use SDL schemas for codegen tests
